feat: sidebar navigation ready
