--- src/shared/components/navigation/SidebarItem.tsx
@@ -1,13 +1,0 @@
-interface Props {
-
-};
-
-function SidebarItem({  }: Props) {
-  return (
-    <div>
-
-    </div>
-  );
-};
-
-export default SidebarItem;
